move non-component exports to separate utils.ts file for pagination
